twitter-bot/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
frozendict==2.4.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
multitasking==0.0.11
numpy==2.2.3
oauthlib==3.2.2
openai==1.65.3
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.6
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
tweepy==4.15.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yfinance==0.2.54
